Backend/backup/requirements.txt
APScheduler==3.10.1
Babel==2.12.1
dateparser==1.2.0
deepmerge==1.1.0
flasgger==0.9.7.1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
flask-marshmallow==0.15.0
flask-mongoengine==1.0.0
Flask-SQLAlchemy==3.1.1
GeoAlchemy2==0.14.2
marshmallow-sqlalchemy==0.29.0
mongoengine==0.27.0
nested-lookup==0.2.25
psycopg2==2.9.6
pymongo==4.3.3
PyPDF2==3.0.1
reportlab==4.0.7
shapely==2.0.1
